Added ptches in migration

+++ db/migrate/20150720232122_remove_price_from_generic_family.rb
@@ -1,10 +1,18 @@
 class RemovePriceFromGenericFamily < ActiveRecord::Migration
   def change
-    remove_column :generic_families, :entrance_price_cents, :integer
-    remove_column :generic_families, :entrance_price_currency, :string
+  	if column_exists?(:generic_families, :entrance_price_cents)
+    	remove_column :generic_families, :entrance_price_cents, :integer
+	end
+  	if column_exists?(:generic_families, :entrance_price_currency)
+    	remove_column :generic_families, :entrance_price_currency, :string
+    end
 
-    remove_column :generic_families, :departure_price_cents, :integer
-    remove_column :generic_families, :departure_price_currency, :string
+    if column_exists?(:generic_families, :departure_price_cents)
+    	remove_column :generic_families, :departure_price_cents, :integer
+    end
+    if column_exists?(:generic_families, :departure_price_currency)
+    	remove_column :generic_families, :departure_price_currency, :string
+    end
     # 
     # remove_money :generic_families, :entrance_price, amount: { null: true, default: nil }, currency: 'USD'
     # remove_money :generic_families, :departure_price, amount: { null: true, default: nil }, currency: 'USD'
